requirements.txt
streamlit
numpy
pillow
matplotlib
scikit-learn
tensorflow
h5py
gdown
